+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.5.9] 2025-05-31
+
 - Add support for additional headers to all endpoints
 - Fix typo in class name `AzureBlob::ForbidenError` to `AzureBlob::ForbiddenError`
 - Fix proper URI encoding for keys with special characters like question marks